+++ RestartSettings.Example.json
@@ -1,11 +1,5 @@
 {
-  "RoomId": 119201858,
+  "RoomId": 12345678,
   "MapIndex": 0,
-  "Round": 1,
-  "Ref": [
-    "31415906"
-  ],
-  "PlayerStatus": {
-    "31415906": true
-  }
+  "Round": 1
 }
