--- RestartSettings.json
@@ -1,11 +1,0 @@
-{
-  "RoomId": 119258335,
-  "MapIndex": 1,
-  "Round": "1",
-  "Ref": [
-    "31415906"
-  ],
-  "PlayerStatus": {
-    "31415906": true
-  }
-}
