+++ match.json
@@ -12,8 +12,8 @@
   "id": "H5",
   "private": false,
   "numberOfRuns": 2,
-  "tournament": "PFC S4",
+  "tournament": "PFC S3",
   "teams": [
-    {"name": "31415906"}
+    {"name": "heipizhu"}
   ]
 }
