+++ optionalWords.json
@@ -14,9 +14,5 @@
   "heipizhu": {
     "join": "PFCS3MVP   heipizhu   来了",
     "leave": "PFCS3MVP   heipizhu   走了"
-  },
-  "Noirtel": {
-    "join": "世界上最最最最最可爱的小蛋糕   Noirtel   来了( •̀ ω •́ )❀",
-    "leave": "世界上最最最最最可爱的小蛋糕   Noirtel   走了TAT"
   }
 }
