+++ pool.json
@@ -1,12 +1,4 @@
 [
-  {
-    "code": "rc1",
-    "id": 1158910,
-    "mod": "freemod"
-  },
-  {
-    "code": "rc2",
-    "id": 4223691,
-    "mod": "freemod"
-  }
+  {"code": "sv1", "id": 4974176, "mod": "freemod"},
+  {"code": "rc1", "id": 5063207, "mod": "freemod"}
 ]
